+++ lib/src/annotations/exception.dart
@@ -1,12 +1,9 @@
 part of jaguar.src.annotations;
 
 /// Exception handler class that handles exception in route chain
-abstract class ExceptionHandler<ExceptionType> {
+abstract class ExceptionHandler {
   const ExceptionHandler();
 
   /// Method that is called when there is an exception in route chain
-  FutureOr<Response> onRouteException(
-      Context ctx, ExceptionType e, StackTrace trace);
-
-  Type exceptionType() => ExceptionType;
+  FutureOr<Response> onRouteException(Context ctx, e, StackTrace trace);
 }