+++ test/bootstrap/serve.dart
@@ -1,7 +1,7 @@
 import 'package:http/http.dart' as http;
 import 'package:test/test.dart';
 import 'package:jaguar/jaguar.dart';
-import 'package:jaguar/bootstrap.dart';
+import 'package:jaguar_reflect/jaguar_reflect.dart';
 import 'package:jaguar_client/jaguar_client.dart';
 
 import 'package:dice/dice.dart';